% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraptR.R
\name{ggraptR}
\alias{ggraptR}
\title{Launch ggraptR in the default browser}
\usage{
ggraptR(initialDf = ggplot2::diamonds, ...)
}
\arguments{
\item{initialDf}{initial dataframe to plot}

\item{...}{extra arguments. Possible options are:
initialPlot - a vector of plot names
externalRun - a boolean to run ggraptR in a separate process
any shiny::runApp arguments}
}
\description{
Launch ggraptR in the default browser
}
\details{
See \url{http://github.com/cargomoose/raptR} for documentation and tutorials
}
\examples{
if (interactive()) {
  ggraptR(initialDf='mtcars', initialPlot=c('Scatter', 'Line'), appDir='inst/ggraptR')
}
}
